- RBG back ground config added for Radar Drawer. - C++ conversion for the algorithm for potential performance improvements. - fixed a bug with asigning target numbers. - added a warning message on start up to warn user about low disk space. - aplication will now clean the PostScript and Render folder for optimization upon exit for next operastion.

+++ AIS_Response_Manager/RadarConsole.py
@@ -1,3 +1,4 @@
+# https://www.geeksforgeeks.org/python/python-gui-tkinter/
 from tkinter import *
 import tkinter as tk
 from tkinter import ttk
@@ -37,7 +38,22 @@ class RadarConsole:
         self.update_image()
 
         # ---- UI ----
+
+        # menu
         self.range_var = StringVar(self.window)
+        
+        # window = tk.Menu(root)
+        # window.config(menu=menu)
+
+        # window.add_cascade(label="File", menu=filemenu)
+        # window.add_command(label="New")
+        # window.add_command(label="Open...")
+        # window.add_separator()
+        # window.add_command(label="Exit", command=root.quit)
+
+        # helpmenu = tk.Menu(menu)
+        # menu.add_cascade(label="Help", menu=helpmenu)
+        # helpmenu.add_command(label="About")
 
         # display
         Label(window,
@@ -71,7 +87,7 @@ class RadarConsole:
                text="KILL TERMINAL",
                font=("Consolas", 14),
                bg="red",
-               command=self.detroy(window),
+               command=self.Detroy,
                ).pack(pady=10)
         
         # Create a Combobox widget for option selection
@@ -87,21 +103,35 @@ class RadarConsole:
                     
         self.radar_loop()
 
-    def detroy(self, window):
+    def Detroy(self):
         self.killswitch = True
-        toDelete = [os.path.join(self.BASE_DIR, f"Radar/PostScript/"), os.path.join(self.BASE_DIR, f"Radar/Renders/")]
-        for folder in toDelete:
-            for filename in os.listdir(folder):
-                file_path = os.path.join(folder, filename)
-                try:
-                    if os.path.isfile(file_path) or os.path.islink(file_path):
-                        os.unlink(file_path)
-                    elif os.path.isdir(file_path):
-                        shutil.rmtree(file_path)
-                except Exception as e:
-                    print('Failed to delete %s. Reason: %s' % (file_path, e))
+        print("Clossing program please wait...")
+        time.sleep(2.1) # give self.radarloop time to register killswitch command, to ensure all files are closed before deletion.
+        self.CleanFiles()
             
-        window.destroy
+    def CleanFiles(self):
+        retry = False
+        if self.BASE_DIR.endswith("AIS_Response_Manager"):
+            toDelete = [os.path.join(self.BASE_DIR, f"Radar/PostScript/"), os.path.join(self.BASE_DIR, f"Radar/Renders/")]
+            for folder in toDelete:
+                for filename in os.listdir(folder):
+                    file_path = os.path.join(folder, filename)
+                    try:
+                        if os.path.isfile(file_path) or os.path.islink(file_path):
+                            os.unlink(file_path)
+                        elif os.path.isdir(file_path):
+                            shutil.rmtree(file_path)
+                    except Exception as e:
+                        print('Failed to delete %s. Reason: %s' % (file_path, e))
+                        retry = True
+                        print("Retrying...")
+        else:
+            print("OS Dirrectory seems to be wrong")
+            print("ensure BASE_DIR is setup properly")
+        
+        if retry:
+            time.sleep(1)
+            self.CleanFiles()
 
     def update_label(self):
         self.range_var.set(f"{self.km_range} KM")
@@ -109,8 +139,6 @@ class RadarConsole:
         if os.path.isfile(self.img_compas):
             self.overlay = Image.open(self.img_compas).convert("RGBA")
         
-        
-
     def change_range(self, delta):
         self.km_range += delta
 
@@ -172,6 +200,11 @@ class RadarConsole:
             print(e)
 
     def radar_loop(self):
+        if self.killswitch == True:
+            self.window.destroy()
+            self.plotter.draw.CloseDrawer()
+            return
+        
         start = time.time()
         self.update_label()
 
@@ -184,8 +217,8 @@ class RadarConsole:
         self.update_image()
 
         # run again after 1000 ms (1 second)
-        if self.window.winfo_exists() and self.killswitch == True:
-            self.window.after(1000, self.radar_loop)
+        if self.window.winfo_exists():
+            self.window.after(500, self.radar_loop)
         end = time.time()
         print(f"Loop performance: {end - start}")
